sobe

+++ metodos.hpp
@@ -8,7 +8,7 @@
 long double bissecao(double a, double b, double aux, double epsilon);
 
 // Função que realiza o cálculo da raiz utilizando o método do posicao falsa
-double pf(double a, double b, double aux, double epsilon);
+long double pf(long double a, long double b, double aux, double epsilon);
 
 // Função que realiza o cálculo da raiz utilizando o método de Newton-Raphson
 long double newton(double a, double b, double epsilon);
